+++ flopy4/mf6/components.py
@@ -1,190 +1,15 @@
-# # Demo attrs/XArray object model
-
-# Demonstrate a tentative `attrs`- and `xarray`-based object model,
-# where `attrs` is used to define classes and `xarray` is used as
-# the underlying data store.
-
-
 from abc import ABC
 from datetime import datetime
 from pathlib import Path
-from typing import Any, Literal, Optional, get_origin
+from typing import Literal, Optional
+from attr import Factory, define, field
+from xarray import DataTree
 
 import numpy as np
-from attr import Attribute, Factory, define, field, fields_dict
-from numpy.typing import ArrayLike, NDArray
-from xarray import Dataset, DataTree
+from numpy.typing import NDArray
 
-
-def _to_path(value: Any) -> Optional[Path]:
-    return Path(value) if value else None
-
-
-def _parse_dim_names(shape: str) -> tuple[str, ...]:
-    return tuple(
-        [
-            dim.strip()
-            for dim in shape.strip()
-            .replace("(", "")
-            .replace(")", "")
-            .split(",")
-            if any(dim)
-        ]
-    )
-
-
-def _try_resolve_dim(data: Optional[DataTree], name: str) -> int | str:
-    name = name.strip()
-    if data is None:
-        return name
-    value = data.get(name, None)
-    if value is not None:
-        return value.item()
-    root = data.root
-    paths = [
-        "tdis",
-        "dis",
-        "gwf/dis",
-    ]
-    for path in paths:
-        try:
-            key = f"{path}/{name}"
-            return root[key].item()
-        except:
-            try:
-                return root[path].dims[name]
-            except:
-                pass
-    return name
-
-
-def _try_resolve_shape(data: DataTree, attr: Attribute) -> tuple[int | str]:
-    shape = attr.metadata.get("shape", None)
-    if shape is None:
-        raise ValueError(f"Array {attr.name} missing shape metadata")
-    shape = [_try_resolve_dim(data, dim) for dim in _parse_dim_names(shape)]
-    return shape
-
-
-def _reshape_array(value: ArrayLike, shape: tuple[int]) -> Optional[NDArray]:
-    value = np.array(value)
-    if value.shape == ():
-        return np.full(shape, value.item())
-    elif value.shape != shape:
-        raise ValueError(
-            f"Shape mismatch, got {value.shape}, expected {shape}"
-        )
-    return value
-
-
-def _resolve_array(
-    self, attr: Attribute, value: Optional[ArrayLike]
-) -> Optional[NDArray]:
-    if value is None:
-        return None
-    shape = _try_resolve_shape(self.data, attr)
-    unresolved = [dim for dim in shape if not isinstance(dim, int)]
-    if any(unresolved):
-        raise ValueError(
-            f"Class '{type(self).__name__}' "
-            f"failed to resolve dims: {', '.join(unresolved)}"
-        )
-    return _reshape_array(value, shape)
-
-
-def _bind_tree(self, parent):
-    parent.data = parent.data.assign({self.data.name: self.data})
-    self.data = parent.data[self.data.name]
-    grandparent = getattr(parent, "parent", None)
-    if grandparent is not None:
-        _bind_tree(parent, grandparent)
-
-
-def _init_tree(self, parent=None, **kwargs):
-    cls = type(self)
-    cls_name = cls.__name__.lower()
-    spec = fields_dict(cls)
-    data = Dataset()
-    dims = set()
-
-    # add arrays
-    for name, attr in spec.items():
-        value = kwargs.get(name, attr.default)
-        shape = attr.metadata.get("shape", None)
-        if shape is not None:
-            dim_names = _parse_dim_names(shape)
-            shape = [
-                _try_resolve_dim(parent.data.root if parent else None, dim)
-                for dim in dim_names
-            ]
-            shape = tuple(
-                [
-                    (dim if isinstance(dim, int) else kwargs.get(dim, dim))
-                    for dim in shape
-                ]
-            )
-            unresolved = [dim for dim in shape if not isinstance(dim, int)]
-            if any(unresolved):
-                raise ValueError(
-                    f"Class '{cls_name}' "
-                    f"failed to resolve dims: {', '.join(unresolved)}"
-                )
-            dims.update(dim_names)
-            value = _reshape_array(value, shape)
-            if value.shape == ():
-                raise ValueError(
-                    f"Failed to resolve array '{name}', "
-                    f"make sure these dimensions exist: "
-                    f"{','.join(dims)}"
-                )
-            data[name] = (dim_names, value)
-
-    # add scalars
-    for name, value in spec.items():
-        if name in data or name in dims:
-            continue
-        value = kwargs.get(name, attr.default)
-        data[name] = value
-
-    self.data = DataTree(data, name=cls_name)
-    if parent is not None:
-        self.parent = parent
-        _bind_tree(self, parent)
-
-
-def _setattr(self, attr: Attribute, value: Any):
-    cls = type(self)
-    spec = fields_dict(cls)
-    if attr.name not in spec:
-        raise AttributeError(f"{cls.__name__} has no attribute {attr.name}")
-    if value is None:
-        return
-    self.data[attr.name] = (
-        (
-            _parse_dim_names(attr.metadata["shape"]),
-            _resolve_array(self, attr, value),
-        )
-        if get_origin(attr.type) in [list, np.ndarray]
-        else value
-    )
-    # TODO run validation?
-
-
-def component(cls):
-    spec = fields_dict(cls)
-
-    def _get(self, name):
-        if name in spec:
-            value = self.data.get(name, None)
-            if value is not None:
-                return value
-            value = self.data.dims.get(name, None)
-            if value is not None:
-                return value
-        return super(cls, self).__getattribute__(name)
-
-    cls.__getattribute__ = _get
-    return cls
+from flopy4.component import component, setattr, resolve_array, init_tree
+from flopy4.utils import to_path
 
 
 class Package(ABC):
@@ -201,7 +26,7 @@ class Sim(ABC):
 
 # @component could in theory wrap the @define decorator
 @component
-@define(init=False, slots=False, on_setattr=_setattr)
+@define(init=False, slots=False, on_setattr=setattr)
 class Dis(Package):
     length_units: str = field(
         # store block as metadata then dynamically
@@ -228,27 +53,27 @@ class Dis(Package):
     delr: NDArray[np.floating] = field(
         # we use a converter both to resolve an array shape
         # and check it, handling both conversion/validation
-        converter=_resolve_array,
+        converter=resolve_array,
         default=1.0,
         metadata={"block": "griddata", "shape": "(ncol,)"},
     )
     delc: NDArray[np.floating] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=1.0,
         metadata={"block": "griddata", "shape": "(nrow,)"},
     )
     top: NDArray[np.floating] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=1.0,
         metadata={"block": "griddata", "shape": "(ncol, nrow)"},
     )
     botm: NDArray[np.floating] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=0.0,
         metadata={"block": "griddata", "shape": "(ncol, nrow, nlay)"},
     )
     idomain: Optional[NDArray[np.integer]] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=1,
         metadata={"block": "griddata", "shape": "(ncol, nrow, nlay)"},
     )
@@ -272,7 +97,7 @@ class Dis(Package):
         botm=0.0,
         idomain=1,
     ):
-        _init_tree(
+        init_tree(
             self,
             parent=model,
             length_units=length_units,
@@ -294,10 +119,10 @@ class Dis(Package):
 
 
 @component
-@define(init=False, slots=False, on_setattr=_setattr)
+@define(init=False, slots=False, on_setattr=setattr)
 class Ic(Package):
     strt: NDArray[np.floating] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=1.0,
         metadata={"block": "packagedata", "shape": "(nodes)"},
     )
@@ -316,7 +141,7 @@ class Ic(Package):
         export_array_ascii=False,
         export_array_netcdf=False,
     ):
-        _init_tree(
+        init_tree(
             self,
             parent=model,
             strt=strt,
@@ -326,7 +151,7 @@ class Ic(Package):
 
 
 @component
-@define(init=False, slots=False, on_setattr=_setattr)
+@define(init=False, slots=False, on_setattr=setattr)
 class Oc(Package):
     @define(slots=False)
     class Format:
@@ -346,17 +171,17 @@ class Oc(Package):
         steps: Optional[list[int]] = field(default=None)
 
     budget_file: Optional[Path] = field(
-        converter=_to_path,
+        converter=to_path,
         default=None,
         metadata={"block": "options"},
     )
     budget_csv_file: Optional[Path] = field(
-        converter=_to_path,
+        converter=to_path,
         default=None,
         metadata={"block": "options"},
     )
     head_file: Optional[Path] = field(
-        converter=_to_path,
+        converter=to_path,
         default=None,
         metadata={"block": "options"},
     )
@@ -377,7 +202,7 @@ class Oc(Package):
         printhead=None,
         perioddata=None,
     ):
-        _init_tree(
+        init_tree(
             self,
             parent=model,
             budget_file=budget_file,
@@ -389,46 +214,46 @@ class Oc(Package):
 
 
 @component
-@define(init=False, slots=False, on_setattr=_setattr)
+@define(init=False, slots=False, on_setattr=setattr)
 class Npf(Package):
     # no options, just arrays for now
     icelltype: NDArray[np.integer] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=0,
         metadata={"block": "griddata", "shape": "(nodes)"},
     )
     k: NDArray[np.floating] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=1.0,
         metadata={"block": "griddata", "shape": "(nodes)"},
     )
     k22: Optional[NDArray[np.floating]] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=None,
         metadata={"block": "griddata", "shape": "(nodes)"},
     )
     k33: Optional[NDArray[np.floating]] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=None,
         metadata={"block": "griddata", "shape": "(nodes)"},
     )
     angle1: Optional[NDArray[np.floating]] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=None,
         metadata={"block": "griddata", "shape": "(nodes)"},
     )
     angle2: Optional[NDArray[np.floating]] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=None,
         metadata={"block": "griddata", "shape": "(nodes)"},
     )
     angle3: Optional[NDArray[np.floating]] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=None,
         metadata={"block": "griddata", "shape": "(nodes)"},
     )
     wetdry: Optional[NDArray[np.floating]] = field(
-        converter=_resolve_array,
+        converter=resolve_array,
         default=None,
         metadata={"block": "griddata", "shape": "(nodes)"},
     )
@@ -445,7 +270,7 @@ class Npf(Package):
         angle3=None,
         wetdry=None,
     ):
-        _init_tree(
+        init_tree(
             self,
             parent=model,
             icelltype=icelltype,
@@ -466,11 +291,11 @@ class Gwf(Model):
         self,
         sim=None,
     ):
-        _init_tree(self, parent=sim)
+        init_tree(self, parent=sim)
 
 
 @component
-@define(init=False, slots=False, on_setattr=_setattr)
+@define(init=False, slots=False, on_setattr=setattr)
 class Tdis(Package):
     @define(slots=False)
     class PeriodData:
@@ -498,7 +323,7 @@ class Tdis(Package):
         time_units=None,
         start_date_time=None,
     ):
-        _init_tree(
+        init_tree(
             self,
             parent=sim,
             nper=nper,
@@ -512,18 +337,4 @@ class Tdis(Package):
 @define(slots=False)
 class Simulation(Sim):
     def __init__(self):
-        _init_tree(self)
-
-
-# Create a simulation.
-
-sim = Simulation()
-tdis = Tdis(sim=sim, nper=1, perioddata=[Tdis.PeriodData()])
-gwf = Gwf(sim=sim)
-dis = Dis(model=gwf)
-ic = Ic(model=gwf, strt=1.0)
-oc = Oc(model=gwf, perioddata=[Oc.Steps()])
-npf = Npf(model=gwf, icelltype=0, k=1.0)
-
-# View the data tree.
-sim.data
+        init_tree(self)
